requirements.txt

Flask==3.0.0
SQLAlchemy<2.0
Flask-SQLAlchemy
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai
